+++ src/overload_numpy/implementors/dispatch.py
@@ -21,8 +21,8 @@ if TYPE_CHECKING:
     import functools
 
     # LOCAL
-    from overload_numpy.wrapper.func import AssistsFunc, ImplementsFunc
-    from overload_numpy.wrapper.ufunc import AssistsUFunc, ImplementsUFunc
+    from overload_numpy.implementors.func import AssistsFunc, ImplementsFunc
+    from overload_numpy.implementors.ufunc import AssistsUFunc, ImplementsUFunc
 
 __all__: list[str] = []
