+++ tests/unittests/wrapper/test_ufunc.py
@@ -6,8 +6,8 @@ import pytest
 
 # LOCAL
 from .test_base import OverloadDecoratorBase_Test, OverloadWrapperBase_Test
-from overload_numpy.wrapper.dispatch import DispatchWrapper
-from overload_numpy.wrapper.ufunc import OverloadUFuncDecorator
+from overload_numpy.implementors.dispatch import DispatchWrapper
+from overload_numpy.implementors.ufunc import OverloadUFuncDecorator
 
 ##############################################################################
 # TESTS
